requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
python-dotenv
elevenlabs
